Refactor retry handler
